Recipe controller views

+++ db/migrate/20220131150536_create_recipes.rb
@@ -2,8 +2,8 @@ class CreateRecipes < ActiveRecord::Migration[7.0]
   def change
     create_table :recipes do |t|
       t.string :name
-      t.integer :prep_time
-      t.integer :cook_time
+      t.decimal :prep_time
+      t.decimal :cook_time
       t.text :description
       t.boolean :public
